Change fileLoc to be 64-bit to allow larger values and avoid borking the value if saving levels

+++ tools/library/tbc/lddecodemetadata.h
@@ -33,7 +33,6 @@
 #include <QDebug>
 #include <array>
 
-#include "vbidecoder.h"
 #include "dropouts.h"
 
 class JsonReader;
@@ -203,7 +202,7 @@ public:
         bool pad = false;
 
         qint32 diskLoc = -1;
-        qint32 fileLoc = -1;
+        qint64 fileLoc = -1;
         qint32 decodeFaults = -1;
         qint32 efmTValues = -1;
